// src/app/modules/admin/model/recogida.model.ts
export interface Recogida {
    id: number;
    contenedor: string;
    titulo: string;
    fecha: string;
    descripcion: string;
    imagen_selector: string;
    imagen_attr: string;
    enlace_selector: string;
    enlace_attr: string;
    url: string;
    fuente: string;
    logo: string;
    activo: boolean;
 }
 